requirements.txt
anthropic 
python-dotenv
PyYAML